refactor(Tokensharp): remove TokenTree-related code and migrate to Trie-based implementation
* Fully removes the `TokenTreeNode` and related builder classes (`TokenTreeNodeBuilder`, `ITokenTreeNode`, `ITokenTreeNodeBuilder`, `TokenTreeNodeExtensions`), deprecating the `TokenTree` structure.
* Refactors namespace usage to adopt `Tokensharp.Trie` instead of `Tokensharp.FastTrie` across the codebase.
* Updates `TokenConfigurationExtensions` to standardize on `TrieNode` construction and drop `TokenTreeNode` references. This simplifies tokenization structures and unifies parsing logic.
